+++ src/factory/accountStatusType.ts
@@ -1,0 +1,14 @@
+/**
+ * 口座ステータスタイプ
+ */
+enum AccountStatusType {
+    /**
+     * 開設済
+     */
+    Opened = 'Opened',
+    /**
+     * 解約済
+     */
+    Closed = 'Closed'
+}
+export default AccountStatusType;